requirements.txt
aiogram==3.3.0
pydantic==2.5.3
pydantic-core==2.14.6
httpx==0.27.2
httpcore==1.0.5
anyio==4.4.0
h11==0.14.0
sniffio==1.3.1
python-dotenv==1.0.1
Pillow==10.4.0
qrcode==7.4.2
uvicorn==0.30.1
fastapi==0.112.0
